Commit new Dto for PointOfInterest Post requests, add validation checks, add controller logic
Adds a new PointOfInterestForCreationDto.cs (leaving out the Id property) to allow Post creation requests by the user. Adds validation checks for the new dto in the form of attributes. Adds logic to PointOfInterestController.cs for creating a new PointOfInterest, including checking of ModelState validity. Adds a model error if the two posted PointOfInterestForCreationDto properties (Id, Name) are the same.