Add and implement Entity Framework, SQLServer nuget packages
Adds EF and SQLServer packages to be used in a code-first approach. Adds a folder 'Entities' to hold objects (including annotations) to be mapped as well as our DbContext. Adds dependency injection to Startup.cs's ConfigureServerices() method to set up the DbContext. Adds DummyController.cs and injects the DbContext, to generate the Db for test purposes.